requirements.txt
streamlit
PyMuPDF
requests
numpy
python-dotenv
loguru
streamlit-pdf-viewer